+++ packages/extension/src/background/index.js
@@ -1,0 +1,7 @@
+export {
+    MessageHost,
+    UnknownMessageTypeError,
+    InvalidMessageError,
+} from './MessageHost.js';
+export { createBackgroundHost } from './createBackgroundHost.js';
+export { attachChromeRuntime } from './ChromeRuntimeAdapter.js';