+++ packages/extension/src/index.js
@@ -14,3 +14,4 @@ export {
     ChromeSessionBackend,
     DEFAULT_SESSION_STORAGE_KEY,
 } from './storage/ChromeSessionBackend.js';
+export * as background from './background/index.js';